requirements.txt
python-telegram-bot
requests
psutil
colorama
rembg
blackboxprotobuf